requirements.txt
psycopg2-binary
openai
tabulate
dotenv
